fixed cursor move bug

+++ src/view.hpp
@@ -35,7 +35,7 @@ namespace View{
 			int getKey(void) const noexcept;
 			void setBK(const short, const short) const;
 			void move(const int, const int) const noexcept;
-			void addString(const std::string&, const short) const noexcept;
+			void addString(const std::string&, const int) const noexcept;
 			void addString(const std::string&) const noexcept;
 			void clear(void) const noexcept;
 			void draw(void) const noexcept;
